+++ AuctionPlatform/Model/Order.cs
@@ -10,7 +10,7 @@ namespace AuctionPlatform.Model
         public string Order_code { get; set; }
         public string Artwork_code { get; set; }
         public string Buyer_id { get; set; }
-        public float Final_price { get; set; }
+        public int Final_price { get; set; }
         public DateTime Order_time { get; set; }
     }
 }
